requirements.txt
streamlit
pytube
openai-whisper
transformers
spacy
torch
ffmpeg-python